requirements.txt
Django==4.2.10
djangorestframework==3.14.0
whitenoise==6.5.0
pillow==10.1.0

# Test deps (optional)
pytest==7.4.0
pytest-django==4.5.2
django-nested-admin
